+++ testament/important_packages.nim
@@ -5,9 +5,8 @@ template pkg(name: string; cmd = "nimble test"; hasDeps = false; url = ""): unty
 var packages*: seq[tuple[name, cmd: string; hasDeps: bool; url: string]] = @[]
 
 
-pkg "NimData", "nim c -o:nimdataa src/nimdata.nim", true
 pkg "argparse"
-pkg "arraymancer", "nim c -r src/arraymancer.nim", true
+pkg "arraymancer", "nim c tests/tests_cpu.nim", true
 pkg "ast_pattern_matching", "nim c -r --oldgensym:on tests/test1.nim"
 pkg "asyncmysql", "", true
 pkg "bigints"
@@ -49,6 +48,7 @@ pkg "neo", "nim c -d:blas=openblas tests/all.nim", true
 pkg "nicy", "nim c src/nicy.nim"
 pkg "nigui", "nim c -o:niguii -r src/nigui.nim"
 pkg "nimcrypto", "nim c -r tests/testall.nim"
+pkg "NimData", "nim c -o:nimdataa src/nimdata.nim", true
 pkg "nimes", "nim c src/nimes.nim", true
 pkg "nimfp", "nim c -o:nfp -r src/fp.nim", true
 pkg "nimgame2", "nim c nimgame2/nimgame.nim", true